# memes_schemas/memes_schemas.py
from schemas.media_schemas import MediaUpdate, MediaRead, MediaCreate


class MemeRead(MediaRead):
    pass


class MemeCreate(MediaCreate):
    pass


class MemeUpdate(MediaUpdate):
    pass